lab1

+++ Lab1/13b.c
@@ -1,3 +1,5 @@
+#include<stdio.h>
+#include<stdlib.h>
 int main(int c, char *v[])
 {
     if (c >= 4 && c <= 21)
@@ -6,9 +8,10 @@ int main(int c, char *v[])
         for (int i = 1; i < c; i++)
         {
 
-            s += atoi(v[i]);
+            s += atof(v[i]);
         }
-        int n = c - 1;
+        // setup cho n để tính tbc
+        int n = c - 1; 
         printf("TBC : %.2f ", s / n);
     }
     else
